### 2023-02-07 #5094 @fboehm
@StefanKohekUMFERI @uellue @suyashkumar @kblin @mdmurbach @kinow @marwahaha - would you be willing to review this submission for the Journal of Open Source Software (JOSS)? JOSS reviews are checklist-driven. You can read more about the review process here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
I'm happy to address any questions that you might have about reviewing for JOSS. Thanks!
