### 2023-02-08 #5135 @editorialbot
[REVIEW]: LaMa: a thematic labelling web application
**Submitting author:** @muctadir (<a href="http://orcid.org/0000-0002-2090-4766">Hossain Muhammad Muctadir</a>)
**Repository:** https://github.com/muctadir/lama
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
