### 2023-02-09 muctadir/lama#5 @kinow
[docs] Add a CONTRIBUTING.md file or similar content to README or docs
Hi,

I am reviewing your JOSS paper :wave: https://github.com/openjournals/joss-reviews/issues/5135

JOSS has a list of requirements for the paper. This issue is about this one:

>Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

It's common for projects hosted at GitHub to contain a `CONTRIBUTING.md` file, e.g. https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors

Your repository doesn't seem to contain that file, and the `README` is also missing that. Maybe that file could contain the instructions for  running & test docs, or the README file could be extended...

Note that the README has a Contributions entry that's confusing too, in the table of contents. Maybe that could be renamed to Acknowledgements, to match the title used.

Cheers,
-Bruno
