### 2023-02-09 muctadir/lama#6 @kinow
[question] Dangling package-lock.json in the root source directory?
Hi,

Part of https://github.com/openjournals/joss-reviews/issues/5135, I checked out the code, and noticed a `package-lock.json` in the source directory. I think that file is normally generated from a `package.json`, and, normally, is stored in the same directory as that file.

Is there a reason for having that file in the root source directory, instead of, having that only in the frontend directory?

Cheers,
-Bruno
