### 2023-02-09 muctadir/lama#7 @kinow
[joss] In the statement of need, how does it compare with OSS annotation tools?
Hi,

Part of https://github.com/openjournals/joss-reviews/issues/5135. I see you mentioned commercial tools in the statement of need of the JOSS paper. The first item in your list of trade offs is the cost. However, that statement of need seems to ignore the existence of other OSS tools that could be compared to LaMa.

Could you consider adding other OSS tools, please? For example:

- https://web.hypothes.is/
- https://github.com/dataqa/nlp-labelling
- https://github.com/heartexlabs/awesome-data-labeling#text
- https://github.com/topics/data-labeling
- https://github.com/topics/text-labeling
- ...
-
Cheers,
-Bruno
