### 2023-02-09 muctadir/lama#9 @kinow
[docs] Improve installation process and instructions for Linux
Hi,

Related to https://github.com/openjournals/joss-reviews/issues/5135; the Linux command in the `README.md` file doesn't work when pasted in the command-line. I've submitted a quick PR for that, see #8

However, running the command I get:

```bash
Please run with privilege
```

Looking at the sources, it runs `docker-compose --env-file .env -f Docker/docker-compose.yml up --build -d`.

If I modify the command, and use `docker compose` instead, that runs fine on my computer without the need of root access (I did that step to run it [rootless](https://docs.docker.com/engine/security/rootless/)).

The script also appears to have some logic to ask the user the operating system,

```bash
echo "What operating system are you running?"
echo "1) Windows"
echo "2) Linux"
read -p 'Boot option: ' os
if [[ $os != 2 ]]
then
    echo "You gave an invalid option for the OS (either not 1/2 or Windows)."
    echo "Try again by reopening the file. Goodbye for now!"
    exit
fi
```

That doesn't look useful as there are two scripts, one of Linux and one for Windows. There could also be MacOS users that could not be sure which version to choose.

IMHO, it'd be easier to just list the docker compose command, and leave the operating system out. Or re-write the installation process to simplify it, and make sure the docs work for users on Linux and/or the possible issues & variations of commands are documented (e.g. maybe say "if you have docker compose, you can run this command... if you have not configured it as rootless you may need to use `sudo`, etc..").

Cheers
-Bruno
