### 2023-02-14 #5135 @muctadir
@kinow thanks for your comments. We noticed the issues that you opened and currently working to resolve them.

@luxaritas Your observation is indeed correct. In terms of commits, my contributions are limited. For example, I improved some implementations as you can see in [https://github.com/muctadir/lama/pull/1](PR#1). However, I contributed significantly at least in two major ways. First is the development of the prior version (as you already mentioned), which includes concepts such as collaborative labeling, conflict resolution, theming and so on. These features were replicated to LaMa although the implementation was done separately. Secondly, I closely supervised the development and influenced all the major decisions. For example, the choice of technologies, moving to service based architecture, the UI design and so on. The newly added features, the changes in architecture, and design are based on our research on the older version of the tool. Please, let me know if my explanation is sufficient.
