### 2023-02-15 d70501f
Updated startup procedure
