### 2023-02-15 muctadir/lama#9 @JarlJansen123
Hi,

I resolved the issues mentioned here, the readme has been updated accordingly. Docker compose commands are listed now rather than making use of shell scripts.

Thanks for the detailed feedback,
Jarl
