### 2023-02-15 muctadir/lama#6 @JarlJansen123
Hi,

The package-lock.json file in source directory was serving no purpose. I have removed it. Now only the functional the package-lock.json in the frontend directory remains.

Thanks for letting us know,
Jarl
