+++ CONTRIBUTING.md
@@ -1,0 +1,12 @@
+# Contributing to LaMa
+
+### Bugs
+If any bugs are detected, please open an issue in the LaMa Github repository (https://github.com/muctadir/lama/issues). If an issue has already been opened about the bug, please use that to provide additional information. In your issue provide a guide on how the bug can be reproduced.
+
+Additionally you can create a pull request to fix any bugs.
+
+### Typos
+Feel free to fix any typos made in the documentation by creating a pull request.
+
+### New features
+Feel free to add new features to the tool by creating a pull request. Before doing so, please create an issue in the LaMa Github repository (https://github.com/muctadir/lama/issues), suggesting the change so that we can have a discussion about its need.
