+++ README.md
@@ -17,7 +17,7 @@ This project builds upon the already existing [Labeling Machine tool](https://gi
 2. [Installation](#installation)
 3. [User Manual](#user_manual)
 4. [Testing](#testing)
-5. [Contributions](#contributions)
+5. [Acknowledgements](#acknowledgements)
 
 ## <a name="dependencies"></a>Dependencies
 
@@ -141,6 +141,6 @@ ng test
 pytest -rA
 ```
 
-## <a name="contributions"></a>Acknowledgements
+## <a name="acknowledgements"></a>Acknowledgements
 
 LaMa was developed based on a web application called [Labeling Machine](https://github.com/muctadir/labeling-machine), which is forked from an earlier version of the [Labeling Machine](https://github.com/emadpres/labeling-machine). Development of LaMa was partially funded by NWO (the Dutch national research council) under the NWO AES Perspective program (Digital Twin), project code P18-03 P3. We acknowledge the contribution of the nine bachelor students from [Eindhoven University of Technology, the Netherlands](https://www.tue.nl) V. Bogachenkova, T. Bradley, E.J. Costa Martins, V. Fürst, B. Henkemans, J.L.O. Jansen, L.C. Lavin, L.V. Nguyen, and A. Olteniceanu, who contributed to this project significantly as part of their final project of the bachelors program in computer science.
