### 2023-02-15 muctadir/lama#5 @JarlJansen123
Hi,

I added a CONTRIBUTING.md file, and changed the readme Contributions to Acknowledgements as suggested.

Thank you for the suggestions,
Jarl
