### 2023-02-16 muctadir/lama#11 @kinow
[joss] Run automated tests with GH Actions?
Hi,

Related to https://github.com/openjournals/joss-reviews/issues/5135, the item

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

Looks like GH Actions is only building the PDF at the moment? Any plans to add tests to the frontend and to the backend, and run them automatically on GH Actions?
