### 2023-02-16 muctadir/lama#11 @kinow
It's not only useful to show that the commits & code are being tested, but also to show users & devs how to run the tests in an automated way.

I followed the README instructions to run the tests, for instance, and it displayed the browser windows for me to close, and then to proceed. I tried `npm run test -- --browsers ChromeHeadless` (I have limited... Angular? knowledge, sorry).
