### 2023-02-16 muctadir/lama#11 @kinow
Another advantage of running in an automated way: I ran the backend tests, and one is failing for me:

```
INFO  [alembic.runtime.migration] Running upgrade  -> 9912972d4f95, empty message
--------------------------------------------------------------------------------------- Captured stderr call ----------------------------------------------------------------------------------------
ERROR [src] Exception on /labelling/create [POST]
Traceback (most recent call last):
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/engine/base.py", line 1819, in _execute_context
    self.dialect.do_execute(
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/engine/default.py", line 732, in do_execute
    cursor.execute(statement, parameters)
sqlite3.IntegrityError: UNIQUE constraint failed: labelling.u_id, labelling.a_id, labelling.lt_id

Traceback (most recent call last):
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/flask/app.py", line 2077, in wsgi_app
    response = self.full_dispatch_request()
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/flask/app.py", line 1525, in full_dispatch_request
    rv = self.handle_user_exception(e)
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/flask_cors/extension.py", line 165, in wrapped_function
    return cors_after_request(app.make_response(f(*args, **kwargs)))
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/flask/app.py", line 1523, in full_dispatch_request
    rv = self.dispatch_request()
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/flask/app.py", line 1509, in dispatch_request
    return self.ensure_sync(self.view_functions[rule.endpoint])(**req.view_args)
  File "/home/kinow/Development/typescript/workspace/lama/backend/src/app_util.py", line 169, in decorated_function
    return f(*args, **kwargs)
  File "/home/kinow/Development/typescript/workspace/lama/backend/src/app_util.py", line 257, in decorated_function
    return f(*args, **kwargs)
  File "/home/kinow/Development/typescript/workspace/lama/backend/src/app_util.py", line 289, in decorated_function
    return f(*args, **kwargs)
  File "/home/kinow/Development/typescript/workspace/lama/backend/src/routes/labelling_routes.py", line 100, in post_labelling
    db.session.commit()
  File "<string>", line 2, in commit
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/orm/session.py", line 1451, in commit
    self._transaction.commit(_to_root=self.future)
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/orm/session.py", line 829, in commit
    self._prepare_impl()
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/orm/session.py", line 808, in _prepare_impl
    self.session.flush()
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/orm/session.py", line 3383, in flush
    self._flush(objects)
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/orm/session.py", line 3522, in _flush
    with util.safe_reraise():
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/util/langhelpers.py", line 70, in __exit__
    compat.raise_(
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/util/compat.py", line 208, in raise_
    raise exception
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/orm/session.py", line 3483, in _flush
    flush_context.execute()
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/orm/unitofwork.py", line 456, in execute
    rec.execute(self)
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/orm/unitofwork.py", line 630, in execute
    util.preloaded.orm_persistence.save_obj(
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/orm/persistence.py", line 245, in save_obj
    _emit_insert_statements(
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/orm/persistence.py", line 1097, in _emit_insert_statements
    c = connection._execute_20(
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/engine/base.py", line 1631, in _execute_20
    return meth(self, args_10style, kwargs_10style, execution_options)
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/sql/elements.py", line 332, in _execute_on_connection
    return connection._execute_clauseelement(
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/engine/base.py", line 1498, in _execute_clauseelement
    ret = self._execute_context(
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/engine/base.py", line 1862, in _execute_context
    self._handle_dbapi_exception(
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/engine/base.py", line 2043, in _handle_dbapi_exception
    util.raise_(
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/util/compat.py", line 208, in raise_
    raise exception
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/engine/base.py", line 1819, in _execute_context
    self.dialect.do_execute(
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/engine/default.py", line 732, in do_execute
    cursor.execute(statement, parameters)
sqlalchemy.exc.IntegrityError: (sqlite3.IntegrityError) UNIQUE constraint failed: labelling.u_id, labelling.a_id, labelling.lt_id
[SQL: INSERT INTO labelling (u_id, a_id, lt_id, l_id, p_id, remark, time) VALUES (?, ?, ?, ?, ?, ?, ?)]
[parameters: (2, 1, 2, 1, 1, 'This is an emotion', '00:00:00.000000')]
(Background on this error at: https://sqlalche.me/e/14/gkpj)
----------------------------------------------------------------------------------------- Captured log call -----------------------------------------------------------------------------------------
ERROR    src:app.py:1455 Exception on /labelling/create [POST]
Traceback (most recent call last):
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/engine/base.py", line 1819, in _execute_context
    self.dialect.do_execute(
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/engine/default.py", line 732, in do_execute
    cursor.execute(statement, parameters)
sqlite3.IntegrityError: UNIQUE constraint failed: labelling.u_id, labelling.a_id, labelling.lt_id

Traceback (most recent call last):
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/flask/app.py", line 2077, in wsgi_app
    response = self.full_dispatch_request()
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/flask/app.py", line 1525, in full_dispatch_request
    rv = self.handle_user_exception(e)
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/flask_cors/extension.py", line 165, in wrapped_function
    return cors_after_request(app.make_response(f(*args, **kwargs)))
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/flask/app.py", line 1523, in full_dispatch_request
    rv = self.dispatch_request()
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/flask/app.py", line 1509, in dispatch_request
    return self.ensure_sync(self.view_functions[rule.endpoint])(**req.view_args)
  File "/home/kinow/Development/typescript/workspace/lama/backend/src/app_util.py", line 169, in decorated_function
    return f(*args, **kwargs)
  File "/home/kinow/Development/typescript/workspace/lama/backend/src/app_util.py", line 257, in decorated_function
    return f(*args, **kwargs)
  File "/home/kinow/Development/typescript/workspace/lama/backend/src/app_util.py", line 289, in decorated_function
    return f(*args, **kwargs)
  File "/home/kinow/Development/typescript/workspace/lama/backend/src/routes/labelling_routes.py", line 100, in post_labelling
    db.session.commit()
  File "<string>", line 2, in commit
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/orm/session.py", line 1451, in commit
    self._transaction.commit(_to_root=self.future)
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/orm/session.py", line 829, in commit
    self._prepare_impl()
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/orm/session.py", line 808, in _prepare_impl
    self.session.flush()
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/orm/session.py", line 3383, in flush
    self._flush(objects)
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/orm/session.py", line 3522, in _flush
    with util.safe_reraise():
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/util/langhelpers.py", line 70, in __exit__
    compat.raise_(
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/util/compat.py", line 208, in raise_
    raise exception
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/orm/session.py", line 3483, in _flush
    flush_context.execute()
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/orm/unitofwork.py", line 456, in execute
    rec.execute(self)
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/orm/unitofwork.py", line 630, in execute
    util.preloaded.orm_persistence.save_obj(
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/orm/persistence.py", line 245, in save_obj
    _emit_insert_statements(
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/orm/persistence.py", line 1097, in _emit_insert_statements
    c = connection._execute_20(
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/engine/base.py", line 1631, in _execute_20
    return meth(self, args_10style, kwargs_10style, execution_options)
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/sql/elements.py", line 332, in _execute_on_connection
    return connection._execute_clauseelement(
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/engine/base.py", line 1498, in _execute_clauseelement
    ret = self._execute_context(
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/engine/base.py", line 1862, in _execute_context
    self._handle_dbapi_exception(
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/engine/base.py", line 2043, in _handle_dbapi_exception
    util.raise_(
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/util/compat.py", line 208, in raise_
    raise exception
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/engine/base.py", line 1819, in _execute_context
    self.dialect.do_execute(
  File "/home/kinow/Development/typescript/workspace/lama/backend/venv/lib/python3.10/site-packages/sqlalchemy/engine/default.py", line 732, in do_execute
    cursor.execute(statement, parameters)
sqlalchemy.exc.IntegrityError: (sqlite3.IntegrityError) UNIQUE constraint failed: labelling.u_id, labelling.a_id, labelling.lt_id
[SQL: INSERT INTO labelling (u_id, a_id, lt_id, l_id, p_id, remark, time) VALUES (?, ?, ?, ?, ?, ?, ?)]
[parameters: (2, 1, 2, 1, 1, 'This is an emotion', '00:00:00.000000')]
(Background on this error at: https://sqlalche.me/e/14/gkpj)
```

Now I have no way of knowing if there's something wrong on my code, if maybe it's trying to access a port bound by the running Docker, or if it's just because I am using the latest code :slightly_smiling_face:
