### 2023-02-16 #5135 @kinow
Hi @muctadir

I've made good progress on the review today. Here's what's pending in my review:

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

Pending, https://github.com/muctadir/lama/issues/11

> State of the field: Do the authors describe how this software compares to other commonly-used packages?

Pending https://github.com/muctadir/lama/issues/7

> Data sharing: If the paper contains original data, data are accessible to the reviewers. If the paper contains no original data, please check this item.
>
> Reproducibility: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.
>
>References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?

The item before is related, about the state of the field, could change the rest of the article, including the references. So I will wait for that one to complete these ones :+1:

That's me for now, ping me again in the issues or here if you have any updates, please.

Thank you!
-Bruno
