### 2023-03-01 muctadir/lama#7 @muctadir
Dear @kinow
Thank you very much for your comments. As we explained in the statement of need, LaMa was developed to aid with the thematic analysis process which is a method for qualitative analysis. Although many of the tools you mentioned are about text annotation, which is a core part of thematic analysis, many of them are ML based (i.e., https://github.com/dataqa/nlp-labelling, https://github.com/BrikerMan/Kashgari). The use-case of LaMa is mostly about manual labeling. Some of the tools you mentioned are about generating/annotating dataset (https://github.com/argilla-io/argilla, https://github.com/RTIInternational/SMART) to be used for different ML algorithms, which is a very different use-case compared to what LaMa tries to solve. The commercial tools that we mentioned in the paper are extremely popular and are developed for qualitative analysis and that is why we mentioned them. At the same time, we did not want this paper to turn into a tool comparison paper and therefore, we left out other text annotation tools that we investigated. Furthermore, the word limitation for the paper contributed to presenting the most important information about the tool itself. Moreover, in the paper, we mentioned 3 points in the statement of need. One of them is indeed cost. However, all the three points equally contributed to our motivation for developing LaMa.
I hope this explains our motivation for the content of the statement of need.
