### 2023-03-01 muctadir/lama#7 @kinow
Hi @muctadir !

What about https://web.hypothes.is/? This is one that I have seen added to Open Source tools to annotate text, and also used in previous companies where I worked. I think that's one of the most popular tools used to annotate text, and  appears to have many overlapping features with Lama.
