+++ backend/src/tests/test_theme.py
@@ -84,10 +84,10 @@ def test_single_theme_info(app, client):
         'labels':
             [{'artifacts':
                 [
-                    {'data': 'Cras pulvinar enim sed posuere molestie. Integer ultrices, leo sit amet hendrerit sodales, diam dui scelerisque nulla, vitae ullamcorper nulla purus ut magna. Cras elementum vitae augue ut maximus. Donec fringilla pellentesque dapibus. Proin non nisl sed sem laoreet luctus a nec leo. Nunc accumsan turpis et orci dictum, vitae efficitur ex ultricies. Nam consectetur, ipsum eget fringilla ultrices, nisi lacus mollis arcu, vitae varius tellus massa eu eros. Pellentesque eget sollicitudin massa, id mollis turpis. Vestibulum in lorem sit amet ligula pulvinar pellentesque eget in leo. Suspendisse blandit, purus non sagittis tempus, sem elit sodales risus, eleifend porttitor metus nunc a ligula. Cras venenatis ex a mi suscipit ornare. In fringilla sit amet ipsum ut molestie. Interdum et malesuada fames ac ante ipsum primis in faucibus.', 'end': None, 'id': 2, 'identifier': '2FA3F', 'p_id': 1, 'parent_id': None, 'start': None},
-                    {'data': 'Pellentesque dui purus, cursus id neque eu, bibendum bibendum nisl. Cras elementum ipsum suscipit pulvinar rhoncus. Duis elementum sit amet leo vitae vehicula. Suspendisse rutrum ut turpis ornare lobortis. Donec pretium, urna eu mollis auctor, arcu sapien ultricies leo, ut tincidunt magna metus eu odio. Cras vel velit felis. Vestibulum congue fermentum augue, in efficitur velit feugiat id. Suspendisse ut mauris sed mauris tempor tincidunt. Aliquam vehicula, diam eget pretium ullamcorper, mi magna hendrerit nibh, at commodo arcu lacus porta erat. Nam quam tortor, vulputate vitae velit quis, lobortis vestibulum tortor. Praesent quam nisi, malesuada eget suscipit sit amet, euismod sit amet arcu. Mauris commodo posuere arcu, sit amet vulputate tortor ultrices iaculis. Vestibulum pretium elit non suscipit commodo. Vestibulum ante ipsum primis in faucibus orci luctus et ultrices posuere cubilia curae;', 'end': None, 'id': 4, 'identifier': '2FA3F', 'p_id': 1, 'parent_id': None, 'start': None},
-                    {'data': 'Vestibulum id condimentum nulla. Nunc dignissim orci ut euismod tempus. Morbi velit felis, tempor sed libero nec, dignissim dignissim turpis. Maecenas non commodo lectus, at elementum est. Sed fermentum interdum odio id eleifend. Quisque suscipit ligula felis, non placerat augue accumsan id. Proin vel velit sem. Aliquam ultrices, enim sit amet feugiat porttitor, purus velit tincidunt ligula, sagittis faucibus nisl massa eu tortor. Quisque vehicula augue lacus, vel iaculis augue posuere id. Morbi eu scelerisque nisi, sit amet pretium nibh. Vivamus tincidunt, ex id fermentum venenatis, justo metus convallis ante, ac auctor augue odio vel massa. Donec tempus est diam, sit amet commodo magna luctus id. In hac habitasse platea dictumst. Proin suscipit ut ligula sit amet pellentesque.', 'end': None, 'id': 8, 'identifier': '2FA3F', 'p_id': 1, 'parent_id': None, 'start': None},
[…]
