### 2023-03-01 f51aa4f
added info when pytest not on path.
