### 2023-03-02 muctadir/lama#7 @muctadir
Dear @kinow

I am looking in the tool that you mentioned and trying to find out the feature set that it provides. For now, I understand that it can annotate texts from various sources and share these annotation across multiple users. I am now trying to find out what other features it provides. Is there a documentation page that you know of? Furthermore, I have a feeling that this tool is about sharing knowledge by something they call "social annotation". And if that is the case, I think the use-case is still different.
Please, let me know what you think about my observation and if I am missing something.
