### 2023-03-05 muctadir/lama#7 @kinow
Hi @muctadir,

I think hypothesis could do with a simpler web page, that lets users find more about its features with less clicks. I had a look and found these resources that should be useful, I hope, for you to view how to get started with it:

- https://web.hypothes.is/help/annotation-basics/
- https://libguides.trinity.edu/digital-annotation/hypothesis
- https://youtu.be/09z5hyBMs8s?t=429 / https://youtu.be/09z5hyBMs8s?t=1067

I think hypothesis has similar features, some interesting features that could be useful in lama (moderation, browser extension), but also it could lack features that are important in lama (like conflict resolution). [This paper](https://www.researchgate.net/publication/319395148_System_Description_KAT_an_Annotation_Tool_for_STEM_Documents), for example, mentions brat and hypothesis, and explains why KAT is still important.

I think the lama paper is doing a great job explaining that there are commercial tools but the complex collaboration is simpler in lama. However, after reading the paper I am still left with the question whether there are Open Source tools that could be used instead (especially important for lama's paper, IMHO, as it's being published in the JOSS).

Hypothesis should check the boxes for Cost and Data access and privacy, but maybe the collaboration workflow doesn't match your use case? Or maybe there are other Open Source annotations tools that have the complex collaboration, but lack the data access and privacy, or tools that have everything that lama does, but are not maintained, etc. I think a short paragraph about it would be enough for the lama paper.

Cheers
Bruno
