### 2023-03-06 muctadir/lama#12 @luxaritas
[ui/ux] Improve display of empty states
In general, when there is no content to display, it is better to show a message explaining this and possibly some additional action for the user to help avoid confusion or an impression that content is failing to load.

Examples:
* When there are no projects, you could include a message like "No projects available" and an inline button to create a new project
* When adding users to a project and none are available, you could include a message like "you're the only user" and an option to invite new users (ie, by email)

Submitted while reviewing openjournals/joss-reviews#5135
