### 2023-03-06 muctadir/lama#13 @luxaritas
[ui/ux] Intermediate account settings page
Currently when going to your account settings, you get a view-only mode and have to additionally click an edit button. Is this necessary? Most sites I'm aware of directly send you to an editable page.

Submitted while reviewing openjournals/joss-reviews#5135
