### 2023-03-06 muctadir/lama#14 @luxaritas
[ui/ux] Navigation persistence
It would probably be more intuitive to have more consistency in the navigation structure. Namely:
- Including a homepage link and account settings on all pages, including while in a project, when editing a project, etc
- Removing the homepage and log out links from the secondary "project" nav (or otherwise have a clearer delineation between navigation items that are specific to the project vs site-wide)
- Instead of redirecting the tagging and conflict pages back to the stats page when not available, either 1) having a page with an empty state (maybe with some indicator of if/how many entries are available to act on, like an inbox) 2) removing them from the sidebar in favor of making them ephemeral pages entered into from the homepage or 3) only show them when there are items available to act on, and visually distinguish them from other nav items (eg changing color, background, etc) so that it's clear they're "exceptional" (reducing confusion around when they are/aren't there and making it clear that something not normally there is now there)

Submitted while reviewing openjournals/joss-reviews#5135
