### 2023-03-06 muctadir/lama#16 @luxaritas
User account provisioning
I imagine in many installations, admins would want to only allow specific individuals to create accounts. It should probably be possible to turn off public account creation in favor of just sending email invites to provision new accounts.

Submitted while reviewing https://github.com/openjournals/joss-reviews/issues/5135
