### 2023-03-06 muctadir/lama#17 @luxaritas
[ui/ux] Improve error messaging
There are a number of instances where obscure errors are presented to the user where it is unclear what the issue was and how the user could correct it. EG:
* When creating a new label with missing fields, an error is just shown saying "invalid input"
* When creating a new account without a description (why would I want to add a description for my own account when creating it?) a 500 error is seen
* When creating a new account with a weak password, just "invalid password" is shown

Related: It should be made clear when fields are required when it isn't obvious, and if a particular format is required (eg specific password validation rules) that should be given up front.

Additionally, errors should ideally be shown directly next to fields they're related to, if applicable

Submitted while reviewing https://github.com/openjournals/joss-reviews/issues/5135
