### 2023-03-06 muctadir/lama#18 @luxaritas
[ui/ux] Introduce max widths
In many areas of the application's layout, it is set such that the contents of the page will go as wide as the page width. particularly when dealing with, say, an ultrawide monitor, this makes it difficult to follow the contents of the page as it hurts things like visual grouping of fields and action buttons and generally hurts the ability to scan the content of the page (having to find small pieces of content in widely separated parts of the screen, reading very short content in very wide fields, text going wider than ~70-80 characters, etc). In many cases, there should likely be maximum widths added either to the entire page layout or specific components of the page (eg, even if you cap the width of all pages, the username and email fields on the account creation screen should be much shorter)

Submitted while reviewing https://github.com/openjournals/joss-reviews/issues/5135
