### 2023-03-06 muctadir/lama#19 @luxaritas
[docs] Improve clarity of installation for development
The docker setup is very straight forward (kudos!) but there doesn't seem to be clear documentation beyond software requirements for getting set up for development (what commands am I expected to run? how can I run things with hot reloading? etc).

Related: Following the instructions provided in the README is not sufficient for running tests successfully. You need to install the python requirements via pip to run the backend tests, and even after installing the frontend requirements via npm and setting my chrome binary, no tests were found.

Submitted while reviewing https://github.com/openjournals/joss-reviews/issues/5135
