### 2023-03-06 muctadir/lama#20 @luxaritas
[joss] Additional detail on prior versions
Currently I'm a little confused about the scope of what is being submitted to JOSS within the context of the previous versions of the application. Namely:
* Is the submission considering this version and the prior two versions as the scope of the contribution?
  * If so, it doesn't look like all prior contributors are included as coauthors, which may suggest a need for revision
  * If not, what is distinct about this version beyond the prior versions? What exactly is the scope of the work done here? What issues or imitations in prior versions were addressed? What is new in this version vs prior versions?

This may warrant some clarification in the paper itself, though I'm not sure if that's necessary (maybe worth checking with the editor?)

Submitted while reviewing https://github.com/openjournals/joss-reviews/issues/5135
