### 2023-03-06 #5135 @luxaritas
Sorry for the delay in review. Overall, this looks great! I've submitted a handful of issues around UI/UX improvements - these aren't intended as blockers to the JOSS review, but are things I hope may be useful to you that I noticed while reviewing.

The following are the items that I feel need resolution:
* https://github.com/muctadir/lama/issues/7 as submitted by @kinow
* https://github.com/muctadir/lama/issues/16 (the lack of access controls here seem potentially problematic considering the stated goals of researcher control over the environment/data)
* https://github.com/muctadir/lama/issues/19 (limitations in contributing guidelines and ability to run automated tests)
* https://github.com/muctadir/lama/issues/20 (clarification on scope and relevant contributors)

Additionally one thing I'm not sure about is the "Example usage" requirement. There are screenshots and such showing generally how to use the application, but I don't think quite "to solve real-world analysis problems". Maybe it could be possible to provide the ability to pre-load synthetic data that shows a realistic use case (not just "lorem ipsum" text)? Haven't submitted an issue for this as I'm not fully clear on what JOSS requires, but I certainly can if that's helpful.
