### 2023-03-06 muctadir/lama#20 @muctadir
> Is the submission considering this version and the prior two versions as the scope of the contribution?

Current version is the scope of this paper. We mentioned https://github.com/emadpres/labeling-machine as this version is the base for developing https://github.com/muctadir/labeling-machine, which in turn influenced the development of the current version, LaMa.

> * If not, what is distinct about this version beyond the prior versions? What exactly is the scope of the work done here? What issues or imitations in prior versions were addressed? What is new in this version vs prior versions?

There are several improvement in the new version. Project management, user management, multi-labeling just to name a few. We only focused on the key features in the paper as explaining all of them would produce a much longer text. For the same reason, we also did not dive deep into any comparison.

> maybe worth checking with the editor?
I am also not sure entirely. Perhaps this is a good idea. I will leave a note at the review thread.
