### 2023-03-06 #5135 @muctadir
Hi @fboehm

@luxaritas has left a comment at https://github.com/muctadir/lama/issues/20. This is about the previous tools LaMa is based upon and how the new version is different. I replied to the comment. However, both of us think that you might want to provide some input there. Would it be possible for you to take a look?

Thanks, Hossain.
