### 2023-03-06 muctadir/lama#7 @muctadir
> However, after reading the paper I am still left with the question whether there are Open Source tools that could be used instead

> I think a short paragraph about it would be enough for the lama paper.

I agree that this is missing. I will add a paragraph based on what we investigated initially before developing LaMa.
