### 2023-03-06 muctadir/lama#20 @fboehm
Thanks for bringing this discussion to my attention, @muctadir. Can you remind me, @muctadir, exactly how you discuss the LaMa predecessors in the paper.md? Please feel free to copy and paste the relevant text of the manuscript in a comment in this thread. Thanks again!
