### 2023-03-06 #5135 @fboehm
@muctadir - I've left a comment on the issue thread, https://github.com/muctadir/lama/issues/20. Please remind me how you currently address the predecessors to LaMa in the paper.md file. My inclination is to include at least a couple sentences to explain the inspirations for LaMa and how LaMa differs from its predecessors. Does that seem reasonable?
