### 2023-03-07 muctadir/lama#20 @muctadir
@fboehm, we mention the predecessors in the final paragraph of the __statement of need__ section of [our paper](https://github.com/openjournals/joss-papers/blob/joss.05135/joss.05135/10.21105.joss.05135.pdf).
