### 2023-03-09 muctadir/lama#16 @JarlJansen123
Interesting suggestion. We are looking at implementing something like this in the next few days.
