+++ backend/requirements.txt
@@ -7,7 +7,7 @@ Flask-Cors==3.0.10
 flask-marshmallow==0.14.0
 Flask-Migrate==3.1.0
 Flask-SQLAlchemy==2.5.1
-greenlet==1.1.2
+greenlet==2.0.2
 iniconfig==1.1.1
 itsdangerous==2.1.2
 Jinja2==3.1.2
