+++ frontend/src/environments/environment.ts
@@ -1,3 +1,4 @@
 export const environment = {
+  production: false,
   apiURL: 'dummy'
 };
