### 2023-03-09 muctadir/lama#19 @JarlJansen123
Sorry about the test-cases not working, I thought I fixed it, but I made a typo somewhere. I fixed my typo. Also I added a more detailed explanation on running the test cases.

For your first point we will provide a detailed description on how to run the tool in a development setting in the next few days. Sorry for the delay.
