### 2023-03-12 muctadir/lama#20 @fboehm
@muctadir - Thank you for the link. Would it be possible to insert a sentence that lists the limitations of the two previous software packages, emadpres/labeling-machine and muctadir/labeling-machine? Maybe this sentence could go immediately before the sentence "Its key features are described in the following section."

I don't feel like a detailed comparision with the labeling-machine software is needed for the manuscript, but I would appreciate the addition of this suggested sentence.
