### 2023-03-16 035eb08
add opensource alternative to resolve #7
