### 2023-03-16 muctadir/lama#7 @muctadir
In light of the current word count in the paper, I have added a sentence to refer to https://labelstud.io/ which we investigated prior to developing LaMa.
