### 2023-03-16 ac348f9
Additional detail on prior versions #20 (#21)
* Additional detail on prior versions #20

* add opensource alternative to resolve #7
