+++ frontend/src/environments/environment.ts
@@ -1,3 +1,5 @@
+// This file with dummy environment variables exists for the tests to run.
+// Actual environment variables are created with setenv.ts (see Readme) and should not be committed
 export const environment = {
   production: false,
   apiURL: 'dummy'
