### 2023-03-16 d551d41
removed typescript + fixed gitignore
the environment.ts file was being tracked for some reason, when it should've been ignored.
