### 2023-03-17 muctadir/lama#7 @kinow
I wasn't aware of Label Studio. Thanks for mentioning it and updating the paper. Looking at this commit, https://github.com/muctadir/lama/commit/ac348f957c17af3477ec5d20992643ecf5dd44eb, the text below lists the cons of the solutions (including Label Studio). The first being "Cost: As these are commercial tools", which is not correct for Label Studio? They appear to have a commercial SASS version, but the code is open source (like LaMa's code, also using a permissive license - ALv2).

The second point is about data access and privacy. Label Studio also has a page about security (https://labelstud.io/guide/security.html) but I think it a wider sense, including database access. But on their documentation you can find more about granting permissions to different users (https://labelstud.io/guide/signup.html#Invite-collaborators-to-a-project). So I think they also provide data access and privacy, and I guess it could be well tested since they are in a commercial operation.

Label Studio also seems to offer extra features like image labelling, ML assisted labelling, and other features related to the third point in the list in the paper, about complex collaboration workflow, e.g.

- relations between annotations - https://labelstud.io/guide/labeling.html#Add-relations-between-annotations
- label with collaborators - https://labelstud.io/guide/labeling.html#Label-with-collaborators

I have not dug into their issues & code, nor signed up for their demo, or tries running it locally. Before doing that, could you elaborate more how it was compared to LaMa, and how did your team identified that it was not sufficient to use Label Studio. Moreover, given that Cost is one of the three items raised as the motivation for LaMa, I think a single Open Source tool is not enough to drive the need for a new tool. It would be better to expand that in the paper too.

We can ping also the editor to have another opinion here, @fboehm, as well the other reviewer @luxaritas
