### 2023-03-20 muctadir/lama#7 @muctadir
> The first being "Cost: As these are commercial tools", which is not correct for Label Studio?

Somehow I missed to fix this text. I just updated the paper with correct text here.

To answer the reminder of the comment I would like to refer to one of my [previous comment](https://github.com/muctadir/lama/issues/7#issuecomment-1449546606). And I would like to quote parts of that reply:

> LaMa was developed to aid with the thematic analysis process which is a method for qualitative analysis. Although many of the tools you mentioned are about text annotation, which is a core part of thematic analysis, many of them are ML based.

> The use-case of LaMa is mostly about manual labeling. Some of the tools you mentioned are about generating/annotating dataset (https://github.com/argilla-io/argilla, https://github.com/RTIInternational/SMART) to be used for different ML algorithms, which is a very different use-case compared to what LaMa tries to solve.

In light of these two comments I made earlier, you can already see how Label Studio has a different use-case, which is about annotating data. You also mentioned about ML assisted labeling with is not what we wanted for LaMa.

> Moreover, in the paper, we mentioned 3 points in the statement of need. One of them is indeed cost. However, all the three points equally contributed to our motivation for developing LaMa.

You focused on cost in your comment and as I mentioned earlier, all the 3 points are equally important. You are indeed, to some extend, correct about the first two points. However, collaboration is also a key motivation which includes features such as collaborative labeling and conflict detection and resolution. To the best of my knowledge, Label Studio does not have such features.

> We can ping also the editor to have another opinion here, @fboehm, as well the other reviewer @luxaritas

I think this might be a good idea.
