### 2023-03-27 muctadir/lama#7 @luxaritas
I haven't spent a _ton_ of time on this, but after looking a little at hypothes.is and label studio, while they're powerful annotation tools, it does not appear to me that they're well suited for thematic analysis, at least in the context of the intended workflow of LaMa. Those tools are all about "pick out a portion of the data that contains some signal" or "classify this piece of data in some existing categories". LaMa however is focused on "we have these pieces of data, and we want to come up with a taxonomy that describes them, coming to a consensus on this taxonomy with other individuals performing the coding". It's a distinctly different type of "annotation" from my understanding of the process.

So, my position here would be that not only may those tools have some deficiencies in the three primary points listed in the paper, they are also likely not suited for the task in general, so there is still an unmet need here.
