### 2023-03-27 #5135 @luxaritas
Hi @fboehm - [my comment above](https://github.com/openjournals/joss-reviews/issues/5135#issuecomment-1455393725) shows pretty well the status from my side - there are still two more issues that I feel are relevant to address. If you could provide some clarity on the requirements around example usage as mentioned there, that could be helpful.
