### 2023-03-27 muctadir/lama#7 @kinow
>it does not appear to me that they're well suited for thematic analysis, at least in the context of the intended workflow of LaMa.
