### 2023-03-29 b249d01
Account activation front-end
Added front-end interface for account activation. Accessible for super-admins only.
