### 2023-03-29 muctadir/lama#7 @fboehm
hi, @luxaritas @kinow and @muctadir - Thanks for the thoughtful discussion here. I especially appreciate the comment from @kinow:
> I believe the paper has been updated to address that cost, but IMHO it would be key to express exactly what you said above. That there are other commercial and open source tools that perform similar tasks, but they lack in handling the complexity of certain annotation workflows, or lack support to controlled ontologies/vocabularies/domains for annotations & labelling, or lack in collaborative data curation, or do not handle thematic analysis, etc. (that, without making the text very long).

Do you all feel that the current version of the manuscript satisfies this request? Thanks again!
