### 2023-03-29 #5135 @fboehm
@luxaritas - Thanks also for your thoughtful comment:

> Additionally one thing I'm not sure about is the "Example usage" requirement. There are screenshots and such showing generally how to use the application, but I don't think quite "to solve real-world analysis problems". Maybe it could be possible to provide the ability to pre-load synthetic data that shows a realistic use case (not just "lorem ipsum" text)? Haven't submitted an issue for this as I'm not fully clear on what JOSS requires, but I certainly can if that's helpful.

I think your suggestion would definitely improve the product. @muctadir - is it possible to implement this suggestion?

Thanks again!
