### 2023-03-29 muctadir/lama#19 @JarlJansen123
A detailed description on how to run the tool for development has been added in the readme.
