### 2023-03-29 muctadir/lama#16 @JarlJansen123
We have implemented something very similar to what you have described here. There is now an environment variable "AUTO_APPROVAL" if this is set to false, then when a user creates a new account it will be pending approval by a super-admin. In other words, there is now control over who can and who can't create an account.

We opted for this implementation over your suggestion using email since we think it is important that this functionality is also usable without access to an email server.

We have provided a short description on the details of how this all works in the section "environment variables" in the readme.
