### 2023-04-04 muctadir/lama#7 @muctadir
Hi @luxaritas @kinow @fboehm
Thanks for your comments. I think it indeed makes sense to be explicit about the use case. I have now updated the paper to include an addition point in the statement of need to address this. To answer @fboehm, I believe, the current version of the paper satisfies the request.
