### 2023-04-04 #5135 @muctadir
@fboehm @kinow @luxaritas
we have addressed all the issues mentioned in https://github.com/openjournals/joss-reviews/issues/5135#issuecomment-1455393725. I believe, the review process can now move forward.

As for the "example usage" pointed out by @fboehm and @luxaritas, to enable this we need to publish some real data with LaMa. However, the data we have at hand are under privacy protected and can not be made publicly available. Furthermore, I also made a quick search to find some open-source thematic analysis/interview data, which did you yield anything useful. Therefore, I am afraid that implementing the suggestion of the "example usage" won't be possible this time around. However, we think this is a very good suggestion and will keep this in mind as a possible enhancement.
