### 2023-04-05 muctadir/lama#7 @kinow
@muctadir I just had a look at the Markdown source and it's looking better! I was trying to preview the PDF, but I think the bot is not updating it. I'll comment in the other issue, preview the PDF, and update this issue & the checklist after that if it's looking OK (from looking at the PDF it was looking fine to me). Cheers
