### 2023-04-05 #5135 @kinow
@fboehm can we get the latest Markdown source of the paper to be re-generated as PDF, please? https://github.com/muctadir/lama/blob/main/paper/paper.md
