### 2023-04-05 muctadir/lama#7 @kinow
@muctadir I thought it would be re-generated by the bot in the pull request. The latest message in the review PR is from Feb 8 (https://github.com/openjournals/joss-reviews/issues/5135#issuecomment-1422935861), but I can't recall if that's how it worked in the past for JOSS reviews, or if I am confusing with another pull request somewhere... will wait for @fboehm 's reply. Thanks!
