### 2023-04-05 muctadir/lama#11 @kinow
Hi @JarlJansen123 , @muctadir ,

I'm going through the review checklist for the JOSS paper again today. Confirmed two items had been fixed.

>Indeed a bit sloppy on our part, sorry for this. I have resolved the issue.

I had a look at the latest GH Actions on this repo, and also at the files under `.github`, but I cannot find any new workflows or automated tests added & executed.

Could you provide some more information on how this has been addressed, please?

Thank you!
