### 2023-04-05 muctadir/lama#11 @kinow
Actually, I believe this is the last point pending in my review now :+1: fixed once we have automated tests that can be executed by readers/users, and that are preferably being run automatically in a CICD pipeline with the code. Cheers.
