### 2023-04-07 muctadir/lama#11 @JarlJansen123
Hi, so in the readme.md there is a detailed description on how to run the tests manually, so automated tests can be executed by readers/users as far as I know. They are currently not automatically executed by a pipeline, but as far as I understand this is not necessary for JOSS submission.

Please let me know if anything is unclear or missing.
