### 2023-04-07 muctadir/lama#11 @kinow
I will try to execute the tests from the readme.

> but as far as I understand this is not necessary for JOSS submission

Correct. You must provide automated tests but they do not need to run in gh actions.

Note, however, that there is a distinction between a Good and an OK approach to pass the review.

https://joss.readthedocs.io/en/latest/review_criteria.html#tests

> Good: An automated test suite hooked up to continuous integration (GitHub Actions, Circle CI, or similar)
>OK: Documented manual steps that can be followed to objectively check the expected functionality of the software (e.g., a sample input file to assert behavior)

I thought since you have gh actions already, hooking up ng and pytest there would be pretty simple.

But if you prefer to keep tests executed manually I think that's Ok too.

If the tests pass on my environment I will check that final item off the review list, and end my review with a note about the tests being manually executed and not with every commit or release (which is Ok to pass the review but I believe it should be made clear to users/readers in the review issue).

Cheers

Bruno
