### 2023-04-07 #5135 @kinow
@fboehm I have re-read the paper now, looked at the code, went through the review checklist, checked everything off and **finished my review**.

The last item was for automated tests. They were added/updated after this issue in the review: https://github.com/muctadir/lama/issues/11

The automated tests are [not being executed in a continuous integration environment, so other devs & users must run it manually](https://github.com/muctadir/lama/issues/11#issuecomment-1500174042) - that's what I did.

Frontend tests:

![Screenshot from 2023-04-07 19-47-32](https://user-images.githubusercontent.com/304786/230654672-89bdd9a1-fdf1-454e-8fc9-94f21f26c9d6.png)

Backend tests:

![Screenshot from 2023-04-07 19-50-21](https://user-images.githubusercontent.com/304786/230654692-5adccb70-ab3d-458a-98af-c671e96cf821.png)

And for the record, the git commit I reviewed and tested:

```bash
(venv) (base) kinow@ranma:~/Development/python/workspace/lama$ git log -n 1
commit 95b4fc6434304952eaa2a85f4d9a25487333d5a1 (HEAD -> main, upstream/main, upstream/HEAD)
Author: Hossain Muctadir <h.m.muctadir@tue.nl>
Date:   Tue Apr 4 08:13:21 2023 +0200

    adjust statement of need
```

Cheers
-Bruno
