### 2023-04-11 #5135 @luxaritas
This is looking great.

@muctadir
> As for the "example usage" pointed out by @fboehm and @luxaritas, to enable this we need to publish some real data with LaMa. However, the data we have at hand are under privacy protected and can not be made publicly available. Furthermore, I also made a quick search to find some open-source thematic analysis/interview data, which did you yield anything useful. Therefore, I am afraid that implementing the suggestion of the "example usage" won't be possible this time around. However, we think this is a very good suggestion and will keep this in mind as a possible enhancement.

Would it be possible to come up with some very simplistic synthetic data to replace the content in the screenshots? I'm imagining something like "I loved the food" labeled as positive and on the topic of food, and "The staff were inattentive" labeled as negative and on the topic of service.
