### 2023-04-11 #5135 @muctadir
> @muctadir - have you had a chance to resolve all of the issues suggested by @luxaritas ? Thanks again!

Hi @fboehm, I believe we resolved all of the blocking issues as indicated by @luxaritas in his https://github.com/openjournals/joss-reviews/issues/5135#issuecomment-1455393725. There are some minor unresolved issues. However, I understand that they are not blocking for the review process?

> Would it be possible to come up with some very simplistic synthetic data to replace the content in the screenshots?

@luxaritas we are looking into it. Thanks.
