### 2023-04-11 #5135 @fboehm
Thanks, @muctadir and @luxaritas. Once @muctadir has a chance to implement the changes suggested above, and quoted here, we can complete the review process.

> Would it be possible to come up with some very simplistic synthetic data to replace the content in the screenshots? I'm imagining something like a project with a topic of restaurant reviews, and artifacts like "I loved the food" labeled as positive and on the topic of food, and "The staff were inattentive" labeled as negative and on the topic of service. That way it's just a little bit more grounded in a concrete use case
