+++ README.md
@@ -123,27 +123,27 @@ Projects are the fundamental building block of the tool, functionalities are on 
 * Join an existing project
 
 Creating a project can be done through the **Create project** button at the top right of the screen, this opens a menu in which settings of the project can be configured.
-![alt text](https://github.com/muctadir/lama/blob/main/images/home-page-user-to-create-project.png?raw=true)
+![alt text](https://github.com/muctadir/lama/blob/main/images/SS1M.png?raw=true)
 
 To join an existing project the project admin of that project needs to add your account through the **project setting menu** of the project. The project settings menu can be opened through the **setting (gear icon)** in the menu bar on the left after opening the project.
 
-![alt text](https://github.com/muctadir/lama/blob/main/images/add-users-button.png?raw=true)
+![alt text](https://github.com/muctadir/lama/blob/main/images/SS2M.png?raw=true)
 
 ### Uploading artifacts
 
 After participating in a project, for any functionality first artifacts have to be added. Artifacts can be added by navigating to the **Artifact management page**, from the navigation menu on the left after opening the project. On the artifact management page click the **Add artifact** button at the top right, and upload a _.txt_ file containing the artifact(s).
 
-![alt text](https://github.com/muctadir/lama/blob/main/images/add-button.png?raw=true)
+![alt text](https://github.com/muctadir/lama/blob/main/images/SS3M.png?raw=true)
 
 ### Labelling an artifact
 
 After having uploaded artifacts it is possible to start labelling them. Navigate using the **navigation menu** on the left to the **labelling page**. An artifact will be shown here. To label an artifact, select labels from the **dropdown menu** below the artifact text.
 
-![alt text](https://github.com/muctadir/lama/blob/main/images/label-type-dropdown.png?raw=true)
+![alt text](https://github.com/muctadir/lama/blob/main/images/SS4M.png?raw=true)
 
 New labels can be created using the **add label** button above the dropdown menu. The remark text-field can be used to add comments on why a certain label was selected.
 
-![alt text](https://github.com/muctadir/lama/blob/main/images/create-new-label.png?raw=true)
+![alt text](https://github.com/muctadir/lama/blob/main/images/SS5M.png?raw=true)
 
 _Note: Labels have a label type, this is a way grouping of labels. To label an artifact the user should select a label from each label type. Label types are specified during the project creation, and are not adjustable afterwards._
 
@@ -151,11 +151,11 @@ _Note: Labels have a label type, this is a way grouping of labels. To label an a
 
 After different users have labelled an artifact there could be conflict between the labels that have been assigned. Artifacts with different labels appear in the **conflict resolution page**. Here users can modify the labelling given to ensure that there is conformity between the assigned labels. 
 
-![alt text](https://github.com/muctadir/lama/blob/main/images/change-label.png?raw=true)
+![alt text](https://github.com/muctadir/lama/blob/main/images/SS6M.png?raw=true)
 
 Clicking the **resolve conflict** button will resolve the conflict if the labels are consistent.
 
-![alt text](https://github.com/muctadir/lama/blob/main/images/resolve-conflict.png?raw=true)
+![alt text](https://github.com/muctadir/lama/blob/main/images/SS7M.png?raw=true)
 
 ### Advanced features
 
