### 2023-04-25 #5135 @fboehm
line 56 "the an" should be "the"
line 70 "artifact based" needs a hyphen, ie, should be "artifact-based"
line 74 "label same" should be "label the same"
line 75 "During labelling " needs a comma, ie, should be "During labelling, "

@muctadir - A few minor corrections are needed, which I list above. The line numbers refer to the pdf document. Please let me know when you've updated the paper.md file to fix these.
