### 2023-04-25 #5135 @fboehm
I've now checked the references and verified that the dois resolve to the specified targets.
