### 2023-04-26 #5135 @muctadir
@fboehm I just finished checking the checklist and the relevant information you can find next to the checked items https://github.com/openjournals/joss-reviews/issues/5135#issuecomment-1522878855
