### 2023-05-15 #5135 @muctadir
@oliviaguest @fboehm
Thanks a lot for your help and support in the publication process. I just noticed that one of the author names are spelled incorrectly in the generated bibtex on the [joss website](https://joss.theoj.org/papers/10.21105/joss.05135). It is spelled Mark __Den van__ Brand, but the correct spelling is Mark __van den__ Brand. I think there might be a mistake while extracting first and last name from the full name in the paper. Is it possible to get it fixed?
